tests/integration-v1/requirements.txt
cattle==0.5.3
gdapi-python==0.5.3
websocket-client==0.23.0
PyJWT==1.4.0

flake8==2.5.1
pytest==3.0.2
pytest-xdist
pyyaml
netaddr
requests
